requirements.txt
flask==3.0.0
streamlit==1.29.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
werkzeug==3.0.1
